Hopefully fixed problem where Thread.sleep() would be interrupted, as well as where 3 emails would be sent daily instead of one. Added a user_blacklist.txt file, where user screen names can be added (in lowercase), signifying the program should ignore tweets form those accounts. Added search_phrases.txt file, which contains the phrases for which twitter should be searched.

+++ src/main/MessageObject.java
@@ -16,7 +16,7 @@ public class MessageObject {
 	public MessageObject(String subject, String text) {
 		this.subject = subject;
 		this.text = text;
-		this.index = -1;
+		this.index = (long) (-1);
 		this.numberOfTweets = -1;
 	}
